move picklists page -> picklist folder
