+++ app/picklist/page.tsx
@@ -5,51 +5,38 @@ import { TeamAvgsI, PickabilityWeightsI } from '@/lib/types/Pickability';
 import { avgDataPipeline } from '@/models/aggregations/averageData';
 import { firstPickWeights, secondPickWeights } from '../../lib/analysis/pickability-weights';
 import { connectToDbB } from '@/middleware/connect-db';
+import {
+    Card,
+    CardContent,
+    CardDescription,
+    CardFooter,
+    CardHeader,
+    CardTitle,
+  } from "@/components/ui/card"
+import { Button } from '@/components/ui/button';
+  
 
-/**
- * Computes pickability using specified weights -- essentially computes a weighted sum
- * @param team - team number
- * @param weights - weights, examples defined in weights.ts
- * @see weights.ts
- *
- * @returns pickability generated using weights
- */
-function computeTeamPickability(team: TeamAvgsI, weights: PickabilityWeightsI) {
-	return Object.keys(weights).reduce((prev: number, key: string | number) => {
-		if (key === 'climbRate') {
-			return prev + (team.climbSuccess * weights.climbRate) / team.forms;
-		} else {
-			return (
-				prev +
-				team[key as keyof PickabilityWeightsI] * weights[key as keyof PickabilityWeightsI]
-			);
-		}
-	}, 0);
-}
+export default async function PicklistPage() {
+    await connectToDbB();
+    const picklists = await PickList.find({});
+    console.log(picklists)
+    return 		<main className='mx-auto max-w-[540px]'>
+   {picklists.map((picklist)=> {
+    return <Card>
+      <Card>
+  <CardHeader>
+    <CardTitle>{picklist.name}</CardTitle>
+    <CardDescription>
+      <p>Updated at: {picklist.updatedAt.toDateString()}</p>
+    </CardDescription>
+  </CardHeader>
+  <CardContent>
+    <Button variant="outline">View</Button>
+  </CardContent>
+</Card>
 
-/**
- * RSC responsible for obtaining initial data and returning the picklist table.
- * @returns page containing DnD picklist
- */
-export default async function PickListPage() {
-	// Connect to DB, get AVGs, sort by first
-	await connectToDbB();
-	const averages = await StandForm.aggregate<TeamAvgsI>(avgDataPipeline).then((res) =>
-		res
-			.map((team) => ({
-				...team,
-				firstPickability: computeTeamPickability(team, firstPickWeights),
-				secondPickability: computeTeamPickability(team, secondPickWeights),
-				climbRate: 100 * (team.climbSuccess / team.forms),
-			}))
-			.sort((b, a) => a.firstPickability - b.firstPickability),
-	);
-	const picklists = (await PickList.find({}))?.map(({ name, ordering }) => ({ name, ordering }));
-
-	return (
-		<main className='mx-auto ml-4 mt-16 p-6'>
-			<h1 className='mb-4 text-primary'>YETI Scouting Picklist</h1>
-			<div>picklist selector interface</div>
-		</main>
-	);
+    </Card>
+    
+   })}
+</main>
 }
